Cat configure.mk and Makefile.shared, strip vars
The recursive make based on Makefile.shared is not worth eliminating, as it's
being used very differently than a typical Makefile: It's being used purely as
a linking tool. Consequently, it has no impact on the dependency graph for the
targets which invoke it.

With the {GNU,BSD}makefile + configure.mk scheme, this change provides an
effective way of ensuring that the configure.mk variables are set without
having to figure out how to negotiate between the different make flavors,
since both flavors accept a Makefile definition streamed into standard input.

This was (mostly) automatically applied by CatConfigureAndMakefileShared() in:
https://code.google.com/p/mike-bland/source/browse/openssl/update_makefiles.py