Local Makefile var names and suffix targets
This is in preparation for including all subdirectory Makefiles in a top-level
Makefile. This will avoid multiply-defined variables that override each other
when the Makefiles are included together. This maintains compatibility with the
existing recursive make structure by replacing the original targets with
recipe-free targets that only depend on the Makefile-local targets.

I compared the build logs from these changes with logs from a clean branch, for
GNU make, BSD make, and GitMake, and they all matched as expected. (GitMake
will now build fips_test_suite regardless, but 'GitMake test' will not execute
it.)

This was automatically applied by UpdateVariableNames(),
EmitSuffixTargetRules(), and UpdateRecursiveMakeArgs() in:
https://code.google.com/p/mike-bland/source/browse/openssl/update_makefiles.py