Update all paths to be relative to the top level
This is part of a series of changes that will result in a broken build until
all of the changes are committed and squashed. The end goal will be a single,
large change that will "flip the switch" to a single-makefile-with-includes
build structure.

This was automatically applied by UpdateDirectoryPaths() in:
https://code.google.com/p/mike-bland/source/browse/openssl/update_makefiles.py