Add implicit rules to GNUmakefiles
Trying to replicate this in BSD syntax appears to be a nonstarter. That, or it
would require an update to the Makefiles to explicitly include all .s and .S
files in particular. That might not be a bad idea, but going to punt for now,
since GNU seems to do the right thing more easily.

This was automatically applied by AddDefaultRules() and
TransformDefaultRuleToGnu() in:
https://code.google.com/p/mike-bland/source/browse/openssl/update_makefiles.py