Extract fips/fips_standalone_sha1 rules
When the single-makefile-with-includes structure kicks in, these rules for the
same object will conflict, though they are mutually exclusive in the recursive
make structure. This change preserves the existing behavior and output; a
future change will make the inclusion of one of the .mk files or other
conditional, to avoid duplicate rules being defined.

Signed-off-by: Mike Bland <mbland@acm.org>